requirements/install.txt
# Requirements for using Asteroid
-r ./torchhub.txt
PyYAML>=5.0
pandas>=0.23.4
pytorch-lightning>=1.5.0
torchaudio>=0.8.0
pb_bss_eval>=0.0.2
torch_stoi>=0.0.1
torch_optimizer>=0.0.1a12,<0.2.0
julius
